Move testing of layer writing into resources folder
