+++ tests/resources/bbox_constants.py
